+++ test/general.test.ts
@@ -45,111 +45,111 @@ describe('test from one table', () => {
   const db = new Database([table], 1)
   const sql = new Builder(db)
 
-  it('Produces [SELECT "col1" FROM testTable;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable";')
   })
 
-  it('Produces [SELECT "col1" AS "C1" FROM testTable;]', () => {
+  it('Produces [SELECT "col1" AS "C1" FROM "testTable";]', () => {
     const actual = sql
       .select(column1.as('C1'))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" AS "C1" FROM testTable;')
+    expect(actual).toEqual('SELECT "col1" AS "C1" FROM "testTable";')
   })
 
-  it('Produces [SELECT "col1" AS "C""1" FROM testTable;] (escape double quote)', () => {
+  it('Produces [SELECT "col1" AS "C""1" FROM "testTable";] (escape double quote)', () => {
     const actual = sql
       .select(column1.as('C"1'))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" AS "C""1" FROM testTable;')
+    expect(actual).toEqual('SELECT "col1" AS "C""1" FROM "testTable";')
   })
 
-  it('Produces [SELECT * FROM testTable;]', () => {
+  it('Produces [SELECT * FROM "testTable";]', () => {
     const actual = sql
       .select(ASTERISK)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT * FROM testTable;')
+    expect(actual).toEqual('SELECT * FROM "testTable";')
   })
 
-  it('Produces [SELECT * FROM testTable;] using selectAsteriskFrom()', () => {
+  it('Produces [SELECT * FROM "testTable";] using selectAsteriskFrom()', () => {
     const actual = sql
       .selectAsteriskFrom(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT * FROM testTable;')
+    expect(actual).toEqual('SELECT * FROM "testTable";')
   })
 
-  it('Produces [SELECT 1 FROM testTable;]', () => {
+  it('Produces [SELECT 1 FROM "testTable";]', () => {
     const actual = sql
       .select(e(1))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT 1 FROM testTable;')
+    expect(actual).toEqual('SELECT 1 FROM "testTable";')
   })
 
-  it('Produces [SELECT 1 AS "One" FROM testTable;]', () => {
+  it('Produces [SELECT 1 AS "One" FROM "testTable";]', () => {
     const actual = sql
       .select(e(1).as('One'))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT 1 AS "One" FROM testTable;')
+    expect(actual).toEqual('SELECT 1 AS "One" FROM "testTable";')
   })
 
-  it("Produces [SELECT 'a' FROM testTable;]", () => {
+  it("Produces [SELECT 'a' FROM \"testTable\";]", () => {
     const actual = sql
       .select(e('a'))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual("SELECT 'a' FROM testTable;")
+    expect(actual).toEqual("SELECT 'a' FROM \"testTable\";")
   })
 
-  it("Produces [SELECT *, NULL, 'a', '*', 1, TRUE, FALSE, -5, 3.14 FROM testTable;]", () => {
+  it("Produces [SELECT *, NULL, 'a', '*', 1, TRUE, FALSE, -5, 3.14 FROM \"testTable\";]", () => {
     const actual = sql
       .select(ASTERISK, null, 'a', '*', 1, true, false, -5, 3.14)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual("SELECT *, NULL, 'a', '*', 1, TRUE, FALSE, -5, 3.14 FROM testTable;")
+    expect(actual).toEqual("SELECT *, NULL, 'a', '*', 1, TRUE, FALSE, -5, 3.14 FROM \"testTable\";")
   })
 
-  it("Produces [SELECT ('a' || 'b') FROM testTable;]", () => {
+  it("Produces [SELECT ('a' || 'b') FROM \"testTable\";]", () => {
     const actual = sql
       .select(e('a', CONCAT, 'b'))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual("SELECT ('a' || 'b') FROM testTable;")
+    expect(actual).toEqual("SELECT ('a' || 'b') FROM \"testTable\";")
   })
 
-  it('Produces [SELECT (1 + (2 - 3)) FROM testTable;]', () => {
+  it('Produces [SELECT (1 + (2 - 3)) FROM "testTable";]', () => {
     const actual = sql
       .select(e(1, ADD, e(2, SUB, 3)))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT (1 + (2 - 3)) FROM testTable;')
+    expect(actual).toEqual('SELECT (1 + (2 - 3)) FROM "testTable";')
   })
 
-  it('Produces [SELECT (1 + (2 - 3)) AS "Calc" FROM testTable;]', () => {
+  it('Produces [SELECT (1 + (2 - 3)) AS "Calc" FROM "testTable";]', () => {
     const actual = sql
       .select(e(1, ADD, e(2, SUB, 3)).as('Calc'))
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT (1 + (2 - 3)) AS "Calc" FROM testTable;')
+    expect(actual).toEqual('SELECT (1 + (2 - 3)) AS "Calc" FROM "testTable";')
   })
 
   describe('select literal values', () => {
@@ -160,46 +160,46 @@ describe('test from one table', () => {
 
   })
 
-  it('Produces [SELECT "col1", "col2" FROM testTable;]', () => {
+  it('Produces [SELECT "col1", "col2" FROM "testTable";]', () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col2" FROM testTable;')
+    expect(actual).toEqual('SELECT "col1", "col2" FROM "testTable";')
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
       .where(column1.eq('x'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x';")
   })
 
-  it('Produces [SELECT "col1" AS "C1", "col2" AS "C2" FROM testTable WHERE "col1" = \'x\';]', () => {
+  it('Produces [SELECT "col1" AS "C1", "col2" AS "C2" FROM "testTable" WHERE "col1" = \'x\';]', () => {
     const actual = sql
       .select(column1.as('C1'), column2.as('C2'))
       .from(table)
       .where(column1.eq('x'))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" AS "C1", "col2" AS "C2" FROM testTable WHERE "col1" = \'x\';')
+    expect(actual).toEqual('SELECT "col1" AS "C1", "col2" AS "C2" FROM "testTable" WHERE "col1" = \'x\';')
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" <> 'x';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" <> 'x';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
       .where(column1.ne('x'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" <> 'x';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" <> 'x';")
   })
 
-  it('Produces [SELECT "col1", "col2" FROM testTable WHERE "col1" = $1;]', () => {
+  it('Produces [SELECT "col1", "col2" FROM "testTable" WHERE "col1" = $1;]', () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -207,14 +207,14 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col2" FROM testTable WHERE "col1" = $1;',
+      sql: 'SELECT "col1", "col2" FROM "testTable" WHERE "col1" = $1;',
       values: ['x'],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col2" FROM testTable WHERE "col1" <> $1;]', () => {
+  it('Produces [SELECT "col1", "col2" FROM "testTable" WHERE "col1" <> $1;]', () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -222,24 +222,24 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col2" FROM testTable WHERE "col1" <> $1;',
+      sql: 'SELECT "col1", "col2" FROM "testTable" WHERE "col1" <> $1;',
       values: ['x'],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" = 5;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" = 5;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
       .where(column4.eq(5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col4" FROM testTable WHERE "col4" = 5;')
+    expect(actual).toEqual('SELECT "col1", "col4" FROM "testTable" WHERE "col4" = 5;')
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" = $1;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" = $1;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
@@ -247,24 +247,24 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col4" FROM testTable WHERE "col4" = $1;',
+      sql: 'SELECT "col1", "col4" FROM "testTable" WHERE "col4" = $1;',
       values: [5],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" <> 5;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" <> 5;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
       .where(column4.ne(5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col4" FROM testTable WHERE "col4" <> 5;')
+    expect(actual).toEqual('SELECT "col1", "col4" FROM "testTable" WHERE "col4" <> 5;')
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" <> $1;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" <> $1;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
@@ -272,33 +272,33 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col4" FROM testTable WHERE "col4" <> $1;',
+      sql: 'SELECT "col1", "col4" FROM "testTable" WHERE "col4" <> $1;',
       values: [5],
     }
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" IS NULL;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NULL;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
       .where(column4.eq(null))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col4" FROM testTable WHERE "col4" IS NULL;')
+    expect(actual).toEqual('SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NULL;')
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" IS NOT NULL;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NOT NULL;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
       .where(column4.ne(null))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col4" FROM testTable WHERE "col4" IS NOT NULL;')
+    expect(actual).toEqual('SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NOT NULL;')
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" IS NOT $1;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NOT $1;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
@@ -306,33 +306,33 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col4" FROM testTable WHERE "col4" IS NOT $1;',
+      sql: 'SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NOT $1;',
       values: [null],
     }
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col1" IS NULL;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col1" IS NULL;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
       .where(column1.eq(null))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col4" FROM testTable WHERE "col1" IS NULL;')
+    expect(actual).toEqual('SELECT "col1", "col4" FROM "testTable" WHERE "col1" IS NULL;')
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col1" IS NOT NULL;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col1" IS NOT NULL;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
       .where(column1.ne(null))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1", "col4" FROM testTable WHERE "col1" IS NOT NULL;')
+    expect(actual).toEqual('SELECT "col1", "col4" FROM "testTable" WHERE "col1" IS NOT NULL;')
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" IS $1;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS $1;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
@@ -340,14 +340,14 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col4" FROM testTable WHERE "col4" IS $1;',
+      sql: 'SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS $1;',
       values: [null],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col4" IS NOT $1;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NOT $1;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
@@ -355,14 +355,14 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col4" FROM testTable WHERE "col4" IS NOT $1;',
+      sql: 'SELECT "col1", "col4" FROM "testTable" WHERE "col4" IS NOT $1;',
       values: [null],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1", "col4" FROM testTable WHERE "col1" IS $1;]', () => {
+  it('Produces [SELECT "col1", "col4" FROM "testTable" WHERE "col1" IS $1;]', () => {
     const actual = sql
       .select(column1, column4)
       .from(table)
@@ -370,24 +370,24 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col4" FROM testTable WHERE "col1" IS $1;',
+      sql: 'SELECT "col1", "col4" FROM "testTable" WHERE "col1" IS $1;',
       values: [null],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' );]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' );]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
       .where(column1.eq('x'), AND, column2.eq('y'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' );")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' );")
   })
 
-  it('Produces [SELECT "col1", "col2" FROM testTable WHERE ( "col1" = $1 AND "col2" = $2 );]', () => {
+  it('Produces [SELECT "col1", "col2" FROM "testTable" WHERE ( "col1" = $1 AND "col2" = $2 );]', () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -395,24 +395,24 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1", "col2" FROM testTable WHERE ( "col1" = $1 AND "col2" = $2 );',
+      sql: 'SELECT "col1", "col2" FROM "testTable" WHERE ( "col1" = $1 AND "col2" = $2 );',
       values: ['x', 'y'],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' );]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' );]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
       .where(column1.eq('x'), OR, column2.eq('y'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' );")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' );")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' AND \"col2\" = 'y';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' AND \"col2\" = 'y';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -420,10 +420,10 @@ describe('test from one table', () => {
       .and(column2.eq('y'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' AND \"col2\" = 'y';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' AND \"col2\" = 'y';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' ) AND \"col3\" = 'z';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' ) AND \"col3\" = 'z';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -431,10 +431,10 @@ describe('test from one table', () => {
       .and(column3.eq('z'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' ) AND \"col3\" = 'z';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' OR \"col2\" = 'y' ) AND \"col3\" = 'z';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' OR \"col2\" = 'y';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' OR \"col2\" = 'y';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -442,10 +442,10 @@ describe('test from one table', () => {
       .or(column2.eq('y'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' OR \"col2\" = 'y';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' OR \"col2\" = 'y';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) OR \"col3\" = 'z';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) OR \"col3\" = 'z';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -453,10 +453,10 @@ describe('test from one table', () => {
       .or(column3.eq('z'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) OR \"col3\" = 'z';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) OR \"col3\" = 'z';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) AND \"col3\" = 'z1' OR \"col3\" = 'z2';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) AND \"col3\" = 'z1' OR \"col3\" = 'z2';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -465,10 +465,10 @@ describe('test from one table', () => {
       .or(column3.eq('z2'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) AND \"col3\" = 'z1' OR \"col3\" = 'z2';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' ) AND \"col3\" = 'z1' OR \"col3\" = 'z2';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' AND \"col2\" = 'y' AND \"col3\" = 'z';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' AND \"col2\" = 'y' AND \"col3\" = 'z';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -477,10 +477,10 @@ describe('test from one table', () => {
       .and(column3.eq('z'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' AND \"col2\" = 'y' AND \"col3\" = 'z';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' AND \"col2\" = 'y' AND \"col3\" = 'z';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' OR \"col2\" = 'y' OR \"col3\" = 'z';]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' OR \"col2\" = 'y' OR \"col3\" = 'z';]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -489,10 +489,10 @@ describe('test from one table', () => {
       .or(column3.eq('z'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x' OR \"col2\" = 'y' OR \"col3\" = 'z';")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x' OR \"col2\" = 'y' OR \"col3\" = 'z';")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x1  x2' AND ( \"col2\" = 'y' OR \"col3\" = 'z' );]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x1  x2' AND ( \"col2\" = 'y' OR \"col3\" = 'z' );]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -500,10 +500,10 @@ describe('test from one table', () => {
       .and(column2.eq('y'), OR, column3.eq('z'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x1  x2' AND ( \"col2\" = 'y' OR \"col3\" = 'z' );")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x1  x2' AND ( \"col2\" = 'y' OR \"col3\" = 'z' );")
   })
 
-  it("Produces [SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x1  x2' OR ( \"col2\" = 'y' AND \"col3\" = 'z' );]", () => {
+  it("Produces [SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x1  x2' OR ( \"col2\" = 'y' AND \"col3\" = 'z' );]", () => {
     const actual = sql
       .select(column1, column2)
       .from(table)
@@ -511,20 +511,20 @@ describe('test from one table', () => {
       .or(column2.eq('y'), AND, column3.eq('z'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM testTable WHERE \"col1\" = 'x1  x2' OR ( \"col2\" = 'y' AND \"col3\" = 'z' );")
+    expect(actual).toEqual("SELECT \"col1\", \"col2\" FROM \"testTable\" WHERE \"col1\" = 'x1  x2' OR ( \"col2\" = 'y' AND \"col3\" = 'z' );")
   })
 
-  it("Produces [SELECT \"col1\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' OR \"col4\" = 5 );]", () => {
+  it("Produces [SELECT \"col1\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' OR \"col4\" = 5 );]", () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column1.eq('x'), AND, column2.eq('y'), OR, column4.eq(5))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\" FROM testTable WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' OR \"col4\" = 5 );")
+    expect(actual).toEqual("SELECT \"col1\" FROM \"testTable\" WHERE ( \"col1\" = 'x' AND \"col2\" = 'y' OR \"col4\" = 5 );")
   })
 
-  it("Produces [SELECT \"col1\" FROM testTable WHERE \"col1\" = 'x' AND ( \"col2\" = 'y' OR \"col3\" = 'z' OR \"col4\" = 5 );]", () => {
+  it("Produces [SELECT \"col1\" FROM \"testTable\" WHERE \"col1\" = 'x' AND ( \"col2\" = 'y' OR \"col3\" = 'z' OR \"col4\" = 5 );]", () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -532,10 +532,10 @@ describe('test from one table', () => {
       .and(column2.eq('y'), OR, column3.eq('z'), OR, column4.eq(5))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\" FROM testTable WHERE \"col1\" = 'x' AND ( \"col2\" = 'y' OR \"col3\" = 'z' OR \"col4\" = 5 );")
+    expect(actual).toEqual("SELECT \"col1\" FROM \"testTable\" WHERE \"col1\" = 'x' AND ( \"col2\" = 'y' OR \"col3\" = 'z' OR \"col4\" = 5 );")
   })
 
-  it("Produces [SELECT \"col1\" FROM testTable WHERE \"col1\" = 'x' OR ( \"col2\" = 'y' AND \"col3\" = 'z' AND \"col4\" = 5 );]", () => {
+  it("Produces [SELECT \"col1\" FROM \"testTable\" WHERE \"col1\" = 'x' OR ( \"col2\" = 'y' AND \"col3\" = 'z' AND \"col4\" = 5 );]", () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -543,10 +543,10 @@ describe('test from one table', () => {
       .or(column2.eq('y'), AND, column3.eq('z'), AND, column4.eq(5))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\" FROM testTable WHERE \"col1\" = 'x' OR ( \"col2\" = 'y' AND \"col3\" = 'z' AND \"col4\" = 5 );")
+    expect(actual).toEqual("SELECT \"col1\" FROM \"testTable\" WHERE \"col1\" = 'x' OR ( \"col2\" = 'y' AND \"col3\" = 'z' AND \"col4\" = 5 );")
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col1" = $1 OR ( "col2" = $2 AND "col3" = $3 AND "col4" = $4 );]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col1" = $1 OR ( "col2" = $2 AND "col3" = $3 AND "col4" = $4 );]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -555,24 +555,24 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col1" = $1 OR ( "col2" = $2 AND "col3" = $3 AND "col4" = $4 );',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col1" = $1 OR ( "col2" = $2 AND "col3" = $3 AND "col4" = $4 );',
       values: ['x', 'y', 'z', 5],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" > 5;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" > 5;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.gt(5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" > 5;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" > 5;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" > $1;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" > $1;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -580,23 +580,23 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col4" > $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col4" > $1;',
       values: [5],
     }
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" < 5;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" < 5;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.lt(5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" < 5;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" < 5;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" < $1;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" < $1;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -604,23 +604,23 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col4" < $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col4" < $1;',
       values: [5],
     }
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" >= 5;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" >= 5;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.ge(5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" >= 5;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" >= 5;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" >= $1;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" >= $1;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -628,23 +628,23 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col4" >= $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col4" >= $1;',
       values: [5],
     }
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" <= 5;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" <= 5;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.le(5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" <= 5;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" <= 5;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" <= $1;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" <= $1;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -652,110 +652,110 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col4" <= $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col4" <= $1;',
       values: [5],
     }
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col1" = "col2";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col1" = "col2";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column1.eq(column2))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col1" = "col2";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col1" = "col2";')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" + "col6");]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" + "col6");]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, ADD, column6))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" + "col6");')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" + "col6");')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" - "col6");]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" - "col6");]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, SUB, column6))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" - "col6");')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" - "col6");')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" - 1);]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" - 1);]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, SUB, 1))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" - 1);')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" - 1);')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" * 1);]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" * 1);]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, MUL, 1))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" * 1);')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" * 1);')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" / 1);]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" / 1);]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, DIV, 1))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" / 1);')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" / 1);')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" % 1);]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" % 1);]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, MOD, 1))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" % 1);')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" % 1);')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = ("col5" ^ 1);]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" ^ 1);]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5, EXP, 1))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = ("col5" ^ 1);')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = ("col5" ^ 1);')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = (1 + "col5");]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = (1 + "col5");]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(1, ADD, column5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = (1 + "col5");')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = (1 + "col5");')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = (1 + 1);]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = (1 + 1);]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(1, ADD, 1))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = (1 + 1);')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = (1 + 1);')
   })
 
   describe('Throw desired Errors', () => {
@@ -786,7 +786,7 @@ describe('test from one table', () => {
         sql.select(column1).from(wrongTable)
       }
 
-      expect(actual).toThrowError('Table: wrongTable not found')
+      expect(actual).toThrowError('Table: "wrongTable" not found')
       expect(actual).toThrowError(TableNotFoundError)
     })
 
@@ -800,37 +800,37 @@ describe('test from one table', () => {
     })
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" > "col5";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" > "col5";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.gt(column5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" > "col5";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" > "col5";')
   })
 
-  it("Produces [SELECT \"col1\" FROM testTable WHERE (\"col7\" > 'tru');]", () => {
+  it("Produces [SELECT \"col1\" FROM \"testTable\" WHERE (\"col7\" > 'tru');]", () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(e(column7, GT, 'tru'))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\" FROM testTable WHERE (\"col7\" > 'tru');")
+    expect(actual).toEqual("SELECT \"col1\" FROM \"testTable\" WHERE (\"col7\" > 'tru');")
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col4" = "col5";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col4" = "col5";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column4.eq(column5))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col4" = "col5";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col4" = "col5";')
   })
 
-  it("Produces [SELECT \"col1\" FROM testTable WHERE \"col2\" = 'value contain single quote '' and more '''' , ''';]", () => {
+  it("Produces [SELECT \"col1\" FROM \"testTable\" WHERE \"col2\" = 'value contain single quote '' and more '''' , ''';]", () => {
     const stringContainSingleQuote = "value contain single quote ' and more '' , '"
     const actual = sql
       .select(column1)
@@ -838,30 +838,30 @@ describe('test from one table', () => {
       .where(column2.eq(stringContainSingleQuote))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\" FROM testTable WHERE \"col2\" = 'value contain single quote '' and more '''' , ''';")
+    expect(actual).toEqual("SELECT \"col1\" FROM \"testTable\" WHERE \"col2\" = 'value contain single quote '' and more '''' , ''';")
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" = TRUE;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" = TRUE;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.eq(true))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7" = TRUE;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7" = TRUE;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" <> TRUE;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" <> TRUE;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.ne(true))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7" <> TRUE;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7" <> TRUE;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" = $1;] for [$1=true]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" = $1;] for [$1=true]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -869,14 +869,14 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col7" = $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col7" = $1;',
       values: [true],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" <> $1;] for [$1=true]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" <> $1;] for [$1=true]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -884,14 +884,14 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col7" <> $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col7" <> $1;',
       values: [true],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" IS $1;] for [$1=null]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" IS $1;] for [$1=null]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -899,14 +899,14 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col7" IS $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col7" IS $1;',
       values: [null],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" IS NOT $1;] for [$1=null]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" IS NOT $1;] for [$1=null]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -914,44 +914,44 @@ describe('test from one table', () => {
       .getPostgresqlBinding()
 
     const expected = {
-      sql: 'SELECT "col1" FROM testTable WHERE "col7" IS NOT $1;',
+      sql: 'SELECT "col1" FROM "testTable" WHERE "col7" IS NOT $1;',
       values: [null],
     }
 
     expect(actual).toEqual(expected)
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7)
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7";')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE NOT "col7";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE NOT "col7";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.not())
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE NOT "col7";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE NOT "col7";')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE (NOT "col7" OR NOT "col8");]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE (NOT "col7" OR NOT "col8");]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.not(), OR, column8.not())
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE ( NOT "col7" OR NOT "col8" );')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE ( NOT "col7" OR NOT "col8" );')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE NOT "col7" AND NOT "col8";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE NOT "col7" AND NOT "col8";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
@@ -959,66 +959,66 @@ describe('test from one table', () => {
       .and(column8.not())
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE NOT "col7" AND NOT "col8";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE NOT "col7" AND NOT "col8";')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" = FALSE;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" = FALSE;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.eq(false))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7" = FALSE;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7" = FALSE;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" = "col8";]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" = "col8";]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.eq(column8))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7" = "col8";')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7" = "col8";')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" IS NULL;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" IS NULL;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.eq(null))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7" IS NULL;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7" IS NULL;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col7" IS NOT NULL;]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col7" IS NOT NULL;]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column7.ne(null))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col7" IS NOT NULL;')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col7" IS NOT NULL;')
   })
 
-  it('Produces [SELECT "col1" FROM testTable WHERE "col1" = ("col2" || "col3");]', () => {
+  it('Produces [SELECT "col1" FROM "testTable" WHERE "col1" = ("col2" || "col3");]', () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column1.eq(column2.concat(column3)))
       .getSQL()
 
-    expect(actual).toEqual('SELECT "col1" FROM testTable WHERE "col1" = ("col2" || "col3");')
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col1" = ("col2" || "col3");')
   })
 
-  it("Produces [SELECT \"col1\" FROM testTable WHERE \"col1\" = (\"col2\" || 'something');]", () => {
+  it("Produces [SELECT \"col1\" FROM \"testTable\" WHERE \"col1\" = (\"col2\" || 'something');]", () => {
     const actual = sql
       .select(column1)
       .from(table)
       .where(column1.eq(column2.concat('something')))
       .getSQL()
 
-    expect(actual).toEqual("SELECT \"col1\" FROM testTable WHERE \"col1\" = (\"col2\" || 'something');")
+    expect(actual).toEqual("SELECT \"col1\" FROM \"testTable\" WHERE \"col1\" = (\"col2\" || 'something');")
   })
 })