+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Fixed
 - ASTERISK, DISTINCT & ALL uses Symbol for better identification
 - BuilderOption refactored by moving it to a new separate file
+- Column name always produced with double quote around them
+- Table name always produced with double quote around them
 
 ## [0.6.1]  2022-01-31
 ### Fixed